+++ MOP/src/GameObjects/Items/ItemHook.cs
@@ -153,7 +153,6 @@ namespace MOP
                 rb.detectCollisions = enabled;
                 rb.isKinematic = !enabled;
                 rb.useGravity = enabled;
-                rb.constraints = enabled ? RigidbodyConstraints.None : RigidbodyConstraints.FreezeAll;
 
                 if (enabled)
                 {