+++ MOP/src/Misc/ExceptionManager.cs
@@ -98,15 +98,17 @@ namespace MOP
 
             output += "=== MOP SETTINGS ===\n\n";
 
-            output += "ActiveDistance: " + MopSettings.ActiveDistance + "\n";
-            output += "ActiveDistanceMultiplicationValue: " + MopSettings.ActiveDistanceMultiplicationValue + "\n";
-            output += "SafeMode: " + MopSettings.SafeMode.ToString() + "\n";
-            output += "RemoveEmptyBeerBottles: " + MopSettings.RemoveEmptyBeerBottles.ToString() + "\n";
-            output += "SatsumaTogglePhysicsOnly: " + MopSettings.SatsumaTogglePhysicsOnly.ToString() + "\n";
-            output += "ToggleVehiclePhysicsOnly: " + MopSettings.ToggleVehiclePhysicsOnly.ToString() + "\n";
-            output += "IgnoreModVehicles: " + MopSettings.IgnoreModVehicles.ToString() + "\n";
-            output += "EnableFramerateLimiter: " + ((bool)MOP.enableFramerateLimiter.GetValue()).ToString() + "\n";
-            output += "FramerateLimiter: " + MOP.framerateLimiter.GetValue().ToString() + "\n"; 
+            output += $"ActiveDistance: {MopSettings.ActiveDistance}\n";
+            output += $"ActiveDistanceMultiplicationValue: {MopSettings.ActiveDistanceMultiplicationValue}\n";
+            output += $"SafeMode: {MopSettings.SafeMode.ToString()}\n";
+            output += $"RemoveEmptyBeerBottles: {MopSettings.RemoveEmptyBeerBottles.ToString()}\n";
+            output += $"SatsumaTogglePhysicsOnly: {MopSettings.SatsumaTogglePhysicsOnly.ToString()}\n";
+            output += $"ToggleVehiclePhysicsOnly: {MopSettings.ToggleVehiclePhysicsOnly.ToString()}\n";
+            output += $"IgnoreModVehicles: {MopSettings.IgnoreModVehicles.ToString()}\n";
+            output += $"EnableFramerateLimiter: {((bool)MOP.enableFramerateLimiter.GetValue()).ToString()}\n";
+            output += $"FramerateLimiter: {MOP.framerateLimiter.GetValue().ToString()}\n";
+            output += $"RulesAutoUpdate: {MOP.rulesAutoUpdate.GetValue().ToString()}\n"; 
+            output += $"RulesAutoUpdateFrequency: {MopSettings.GetRuleFilesUpdateDaysFrequency()}\n"; 
             output += "ExperimentalBranch: " + ModLoader.CheckIfExperimental() + "\n";
 
             // List installed mods.